+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Additions
 Improvements
 - Improved EnumHelper.GetValues signature to return an array
 - Allow using external gesture handling alongside InputHandler through ExternalGestureHandling
+- Discard old data when updating a StaticSpriteBatch
 
 Fixes
 - Fixed TokenizedString handling trailing spaces incorrectly in the last line of non-left aligned text